backend/requirements.txt
# LogAnalyzer - Docker Log Analysis Service

# Web Framework
fastapi>=0.115.0
uvicorn>=0.32.0

# HTTP Client
httpx>=0.27.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Scheduler
apscheduler>=3.10.4

# Docker
docker>=7.1.0

# Config
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# GitHub API
PyGithub>=2.1.1

# Date utilities
python-dateutil>=2.8.2
